Got heap working in LM
Fixed a few bugs related to using the heap in LM

+++ arch/x86/lmode/spinlock.h
@@ -2,7 +2,7 @@
 
 #include <lib/types.h>
 
-typedef volatile uint32_t mutex_t;
+typedef volatile size_t mutex_t;
 
 void spinlock_acquire(mutex_t *lock);
 void spinlock_release(mutex_t *lock);
